Moved Colors tool to under Materials tool
