+++ Arma3ObjectBuilder/ui/__init__.py
@@ -7,7 +7,6 @@ from . import props_action
 from . import props_material
 from . import props_object_mesh
 from . import tool_outliner
-from . import tool_color
 from . import tool_hitpoint
 from . import tool_mass
 from . import tool_materials
